Add the VSvgNode Interface, update renderers accordingly
The base vdom node doesn't have any information about what
kinds of attributes are available. Here we add the VSvgNode, which
is a child of the VDomNode and defines what attributes are available
for Svg-based elements and what style props are available.

The styling attributes were pulled from React-Native's viewProps, and the common svg props were pulled from react-native-svg. The idea with
that is that by going mobile-first, we can adapt to the web platform
pretty quickly, but the inverse may not necessarily be true.