Streamline scale creation
When scales are created, their domain can either be bound to an array of values, a function that emits an array of values, or a bindind-expression.

Previously, we used the binding expression in tandem with the a table property to pull of the databind. However, the table-property was only used in conjenction with a bind expression, so it added a bit of noise to the bind overall.

This replaces the previous binding mechanism of `table="data" domain="field"` with `domain="data.field"`, which cleans up the binding a bit.